Add new Button & Input
