+++ src/components/Input/Input.tsx
@@ -1,55 +1,41 @@
 import React from 'react'
 import styled from 'styled-components'
+import {
+    Input as BaseInput,
+    InputProps as BaseInputProps,
+    Typography,
+    Button,
+} from 'antd'
+const { Text } = Typography
 
-export interface InputProps {
-	endAdornment?: React.ReactNode
-	onChange: (e: React.FormEvent<HTMLInputElement>) => void
-	placeholder?: string
-	startAdornment?: React.ReactNode
-	value: string
+const StyledInput = styled(BaseInput)``
+
+export interface InputProps extends BaseInputProps {
+    onClickMax?: () => void
 }
 
 const Input: React.FC<InputProps> = ({
-	endAdornment,
-	onChange,
-	placeholder,
-	startAdornment,
-	value,
-}) => {
-	return (
-		<StyledInputWrapper>
-			{!!startAdornment && startAdornment}
-			<StyledInput
-				placeholder={placeholder}
-				value={value}
-				onChange={onChange}
-			/>
-			{!!endAdornment && endAdornment}
-		</StyledInputWrapper>
-	)
-}
-
-const StyledInputWrapper = styled.div`
-	align-items: center;
-	// background-color: ${(props) => props.theme.color.grey[200]};
-	border-radius: ${(props) => props.theme.borderRadius}px;
-	// box-shadow: inset 4px 4px 8px ${(props) => props.theme.color.grey[300]},
-	//   inset -6px -6px 12px ${(props) => props.theme.color.grey[100]};
-	box-shadow: inset 0 0 8px ${({ theme }) => theme.color.grey[400]};
-	display: flex;
-	height: 72px;
-	padding: 0 ${(props) => props.theme.spacing[3]}px;
-`
-
-const StyledInput = styled.input`
-	background: none;
-	border: 0;
-	font-size: 18px;
-	flex: 1;
-	height: 56px;
-	margin: 0;
-	padding: 0;
-	outline: none;
-`
+    children,
+    onClickMax,
+    suffix,
+    ...rest
+}) => (
+    <StyledInput
+        suffix={
+            onClickMax ? (
+                <>
+                    {suffix && <Text type="secondary">{suffix}</Text>}
+					&nbsp;
+                    <Button block size="small" onClick={onClickMax}>
+                        MAX
+					</Button>
+                </>
+            ) : (
+                suffix
+            )
+        }
+        {...rest}
+    />
+)
 
 export default Input
